M9A: fix static group layer fixture type

+++ crates/runenui_runtime/tests/m9_static_node_effect_groups.rs
@@ -20,7 +20,7 @@ fn color(red: u8) -> Color {
 
 #[derive(Debug)]
 struct LayeredPaint {
-    entries: &'static [(i32, u8)],
+    entries: &'static [(i64, u8)],
 }
 
 impl Widget<()> for LayeredPaint {
